refactor(frontend): extract ExpandedBody to shared module
Moves ExpandedBody, MetadataRow and FallbackStats from InsightsPage.tsx
into components/insights/ExpandedBody.tsx so the dashboard can reuse the
same expand UI. ExpandedBodyInsight is a minimal prop shape — metric and
computed_at are optional so dashboard's DashboardInsight (no metric field)
passes without issue. InsightsPage now imports from the shared module.
